refactor(sync): promote LinkDeviceHowItWorksCard to standalone widget (#388)
The Link Device screen still carried three private classes inline
(`_ThisDeviceCard`, `_ImportFromDeviceCard`, `_HowItWorksCard`)
after #417 first split them out. Promote the simplest one —
`_HowItWorksCard` — to a standalone widget under
`presentation/widgets/`, since it has zero state dependencies and
the screen file can shrink immediately.

`link_device_screen.dart`: 276 → 232 lines (–44).

Regression tests in `link_device_how_it_works_card_test.dart`
(3 tests):
1. renders the title and the info icon
2. renders the step-by-step body text (Device A / Device B / Import)
3. renders inside a Card

All 105 sync tests still pass.

Refs #388
