+++ lib/features/sync/presentation/screens/link_device_screen.dart
@@ -6,6 +6,7 @@ import '../../../../core/sync/supabase_client.dart';
 import '../../../../core/widgets/snackbar_helper.dart';
 import '../../../../l10n/app_localizations.dart';
 import '../../providers/link_device_provider.dart';
+import '../widgets/link_device_how_it_works_card.dart';
 
 class LinkDeviceScreen extends ConsumerStatefulWidget {
   const LinkDeviceScreen({super.key});
@@ -42,7 +43,7 @@ class _LinkDeviceScreenState extends ConsumerState<LinkDeviceScreen> {
           const SizedBox(height: 16),
           _ImportFromDeviceCard(codeController: _codeController),
           const SizedBox(height: 16),
-          const _HowItWorksCard(),
+          const LinkDeviceHowItWorksCard(),
         ],
       ),
     );
@@ -229,48 +230,3 @@ class _ImportFromDeviceCard extends ConsumerWidget {
   }
 }
 
-/// Explanation card describing how the merge works step-by-step.
-class _HowItWorksCard extends StatelessWidget {
-  const _HowItWorksCard();
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final l10n = AppLocalizations.of(context);
-
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                const Icon(Icons.info_outline, size: 20),
-                const SizedBox(width: 8),
-                Text(
-                  l10n?.linkDeviceHowItWorksTitle ?? 'How it works',
-                  style: theme.textTheme.titleSmall
-                      ?.copyWith(fontWeight: FontWeight.bold),
-                ),
-              ],
-            ),
-            const SizedBox(height: 8),
-            Text(
-              l10n?.linkDeviceHowItWorksBody ??
-                  '1. On Device A: copy the device code above\n'
-                      '2. On Device B: paste it in the "Device code" field\n'
-                      '3. Tap "Import data" to merge favorites and alerts\n'
-                      '4. Both devices will have all combined data\n\n'
-                      'Each device keeps its own anonymous identity. '
-                      'Data is merged, not moved.',
-              style: theme.textTheme.bodySmall?.copyWith(
-                color: theme.colorScheme.onSurfaceVariant,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
